+++ templates/panier/index.html.twig
@@ -2,13 +2,16 @@
 
 {% block title %}Panier - BD Commerce
 {% endblock %}
+{% block js %}
+	<script src="https://js.stripe.com/v3/"></script>
+{% endblock %}
 
 {% block content %}
 
 	{% if panier is defined %}
 		{# <pre>
-																												    {{ dump(panier) }}
-																												</pre> #}
+																																																																																										    {{ dump(panier) }}
+																																																																																										</pre> #}
 
 		<table class="table">
 			<thead>
@@ -58,7 +61,9 @@
 				articles):
 				<b>{{total}}€</b>
 			</p>
-			<a href="{{ path('commande') }}" class="btn btn-success mb-1">Commander</a><br>
+
+			{# <a href="{{ path('stripe_create_session') }}" class="btn btn-success mb-1">Commander</a><br> #}
+			<a class="btn btn-success mb-1" id="checkout-button">Commander</a><br>
 			<a href="{{ path('remove_panier') }}" class="btn btn-danger">Vider le panier</a>
 		</div>
 	{% else %}
@@ -66,4 +71,48 @@
 			Vous n'avez pas de BD dans votre panier
 		</p>
 	{% endif %}
+{% endblock %}
+
+{% block script %}
+	<script type="text/javascript">
+
+		// Create an instance of the Stripe object with your publishable API key
+
+var stripe = Stripe("pk_test_51IY7XADH5SafTNx8JecDMTpHLPhtD6XiWnB6mmLoJ8iczxB8FFjwRhuB5p3J9RXep9v7Q2zFNA2oMVLUutQV7JHD00Zarap09x");
+
+var checkoutButton = document.getElementById("checkout-button");
+
+checkoutButton.addEventListener("click", function () {
+
+fetch("/commande/create-session", {method: "POST"}).then(function (response) {
+
+return response.json();
+
+}).then(function (session) {
+
+return stripe.redirectToCheckout({sessionId: session.id});
+
+}).then(function (result) {
+
+// If redirectToCheckout fails due to a browser or network
+
+// error, you should display the localized error message to your
+
+// customer using error.message.
+
+if (result.error) {
+
+alert(result.error.message);
+
+}
+
+}).catch(function (error) {
+
+console.error("MF ! Error:", error);
+
+});
+
+});
+
+	</script>
 {% endblock %}
